move Level.kt to :shared:domain
