+++ Android/domain/src/main/java/co/kr/searchvoca/domain/model/Word.kt
@@ -1,5 +1,6 @@
 package co.kr.searchvoca.domain.model
 
+import co.kr.searchvoca.shared.domain.Level
 import java.io.Serializable
 
 data class Word(
