+++ Android/shared/domain/src/main/java/co/kr/searchvoca/shared/domain/Level.kt
@@ -1,4 +1,4 @@
-package co.kr.searchvoca.domain.model
+package co.kr.searchvoca.shared.domain
 
 enum class Level {
     EASY,
